+++ torii-example/Changelog.md
@@ -5,6 +5,7 @@
 ### v0.0.0-dev7
 
 - Add support to for the "folder" component.
+- Add project sidebar quick menu: New File, Refresh Files
 - Rewire the autocomplete feature to see cached files. (You need to open the folder in the file tree for it to be visible by autocompletion)
 - A lot of client-side refactoring, rewrite the project sidebar with custom tree implementation.